requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
psycopg2-binary
pgvector
langchain
python-dotenv
uvicorn
daphne
requests
beautifulsoup4
langchain-text-splitters
boto3
langchain-aws
celery
django-celery-beat
redis
